fixed extra spaces and console.log calls; fixed routes folder structure; fixed file names accoring to their parent folders
