+++ server/configs/colors.config.js
