+++ server/helpers/provinces.helper.js
@@ -1,20 +1,15 @@
 var request = require('request'),
-	urlsConfig = require('../configs/urls'),
+    urlsConfig = require('../configs/urls.config'),
 	url = require('url'),
 	vow = require('vow');
 
-var request = require('request');
-
-function getTemp(array) {
+function getProvinces(geoid) {
 	var deferred = vow.defer();
 
 	var uri = url.format({
 	    protocol: 'http',
 	    hostname: urlsConfig.api,
-	    pathname: 'factual',
-	    query: {
-	        ids: array.join(',')
-	    }
+	    pathname: ['localities', geoid, 'provinces'].join('/')
 	});
 
 	request.get({
@@ -31,4 +26,4 @@ function getTemp(array) {
 	return deferred.promise();
 }
 
-module.exports = getTemp;
+module.exports = getProvinces;
