+++ server/locales/common.locale.js
@@ -1,0 +1,3 @@
+module.exports = {
+    pageTitle: 'Yandex Weather {cityid}'
+};
