+++ static/pages/views/forecast.full.item.row.jsx
@@ -1,9 +1,8 @@
 /** @jsx React.DOM */
-var timeOfDay = require('../../../server/locales/timeofday.js');
+var timeOfDay = require('../../../server/locales/timeofday.locale.js');
 
 var ForecastFullItemRow = React.createClass({
     render: function() {
-        console.log(timeOfDay);
         var item = this.props.part;
         return (
             <div className="forecast-full__weather__item" style={{"background-color": item.color}}>
